Base: Use the correct memory unit suffixes in `about::processes`
Previously the GB suffix was displayed as BB.
